+++ src/main/java/com/university/crud/springbootcrudrest/messaging/StudentJmsSender.java
@@ -1,4 +1,4 @@
-package com.university.crud.springbootcrudrest.service;
+package com.university.crud.springbootcrudrest.messaging;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jms.core.JmsTemplate;
